refactor: 게시글 목록 조회 dto 이름 수정
