--- src/main/java/com/ssafy/ssafsound/domain/post/dto/GetPostListResDto.java
@@ -1,12 +1,0 @@
-package com.ssafy.ssafsound.domain.post.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@Builder
-public class GetPostListResDto {
-    private List<GetPostResDto> posts;
-}
